Allow to define confdir
PAM has a pam_start_confdir() which allows to define the configuration
directory where all services are located.
This is useful to define your own service on tests in particular, so
that you can control your stack and be independant of the host when
running them.
Allow defining this configuration directory, with functional options.
This keeps then the current package API while avoiding proliferation of
function namse, which would be doubled due to Start() and StartFunc()
flavors.